Rename OpenCVWrapper to SpectatorViewPlugin
