+++ SpectatorViewPlugin/SpectatorViewPlugin/SpectatorViewPlugin/MarkerDetector.cpp
