+++ SpectatorViewPlugin/SpectatorViewPlugin/SpectatorViewPlugin/MarkerDetector.h
