+++ SpectatorViewPlugin/SpectatorViewPlugin/SpectatorViewPlugin/dllmain.cpp
